Abstracting common code in file upload actions
