+++ components/box/actions/upload-file/upload-file.mjs
@@ -1,7 +1,5 @@
 import app from "../../box.app.mjs";
-import { ConfigurationError } from "@pipedream/platform";
-import FormData from "form-data";
-import utils from "../../common/utils.mjs";
+import { getFileUploadBody } from "../../common/common-file-upload.mjs";
 
 export default {
   name: "Upload a File",
@@ -42,30 +40,19 @@ export default {
       ],
     },
   },
+  methods: {
+    getFileUploadBody,
+  },
   async run({ $ }) {
-    const fileValidation = utils.isValidFile(this.file);
-    if (!fileValidation) {
-      throw new ConfigurationError("`file` must be a valid file path!");
-    }
-    const fileMeta = utils.getFileMeta(fileValidation);
-    const fileContent = utils.getFileStream(fileValidation);
-    const attributes = fileMeta.attributes;
-    if (this.createdAt && utils.checkRFC3339(this.createdAt)) {
-      attributes.content_created_at = this.createdAt;
-    }
-    if (this.modifiedAt && utils.checkRFC3339(this.modifiedAt)) {
-      attributes.content_modified_at = this.modifiedAt;
-    }
-    if (this.fileName) {
-      attributes.name = this.fileName;
-    }
-    if (this.parentId) {
-      attributes.parent.id = this.parentId;
-    }
-    const data = new FormData();
-    data.append("attributes", JSON.stringify(attributes));
-    data.append("file", fileContent, {
-      knownLength: fileMeta.size,
+    const {
+      file, createdAt, modifiedAt, fileName, parentId,
+    } = this;
+    const data = this.getFileUploadBody({
+      file,
+      createdAt,
+      modifiedAt,
+      fileName,
+      parentId,
     });
     const response = await this.app.uploadFile({
       $,
